Use int for digital sensor values.

+++ src/CayenneArduinoMQTTClient.h
@@ -283,7 +283,7 @@ public:
 	* @param channel  Cayenne channel number
 	* @param value  Value to be sent
 	*/
-	void digitalSensorWrite(unsigned int channel, float value)
+	void digitalSensorWrite(unsigned int channel, int value)
 	{
 		virtualWrite(channel, value, CAYENNE_FLASH(TYPE_DIGITAL_SENSOR), CAYENNE_FLASH(UNIT_DIGITAL));
 	}
